+++ services/hizone-interaction/src/main/java/com/example/service/InteractionService.java
@@ -2,14 +2,14 @@ package com.example.service;
 
 import java.util.List;
 
-import com.example.hizone.dao.interaction.Interaction;
-import com.example.hizone.front.interaction.CancelCollectPost;
-import com.example.hizone.front.interaction.CancelLikePost;
-import com.example.hizone.front.interaction.CollectPost;
-import com.example.hizone.front.interaction.LikePost;
-import com.example.hizone.inter.UpdateCommentCount;
-import com.example.hizone.inter.UserInteraction;
-import com.example.hizone.inter.UserPost;
+import com.example.hizone.dto.UpdateCommentCount;
+import com.example.hizone.dto.UserInteraction;
+import com.example.hizone.dto.UserPost;
+import com.example.hizone.request.interaction.CancelCollectPost;
+import com.example.hizone.request.interaction.CancelLikePost;
+import com.example.hizone.request.interaction.CollectPost;
+import com.example.hizone.request.interaction.LikePost;
+import com.example.hizone.table.interaction.Interaction;
 
 public interface InteractionService {
 
@@ -17,23 +17,23 @@ public interface InteractionService {
 
     void addCollectPost(CollectPost collectPost);
 
-    Interaction getInteraction(int postId);
+    Interaction getInteraction(Long postId);
 
-    void initInteraction(int postId);
+    void initInteraction(Long postId);
 
     UserInteraction getUserInteraction(UserPost userPost);
 
-    List<LikePost> getLikePostList(int postId);
+    List<LikePost> getLikePostList(Long postId);
 
-    List<CollectPost> getCollectPostList(int postId);
+    List<CollectPost> getCollectPostList(Long postId);
 
     void cancelLikePost(CancelLikePost cancelLikePost);
 
     void cancelCollectPost(CancelCollectPost cancelCollectPost);
 
-    void updatePostLikeCount(int postId, int increment);
+    void updatePostLikeCount(Long postId, Long increment);
 
-    void updatePostCollectCount(int postId, int increment);
+    void updatePostCollectCount(Long postId, Long increment);
 
     void updateCommentCount(UpdateCommentCount updateCommentCount);
 }
