+++ services/hizone-user/src/main/java/com/example/mapper/UserMapper.java
@@ -6,23 +6,23 @@ import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Select;
 import org.apache.ibatis.annotations.Update;
 
-import com.example.hizone.dao.user.User;
-import com.example.hizone.dao.user.UserMetadata;
-import com.example.hizone.front.user.UpdateUserInfo;
-import com.example.hizone.inter.UpdateUserMetadata;
-import com.example.hizone.outer.UserDetail;
+import com.example.hizone.dto.UpdateUserMetadata;
+import com.example.hizone.request.user.UpdateUserInfo;
+import com.example.hizone.response.UserDetail;
+import com.example.hizone.table.user.User;
+import com.example.hizone.table.user.UserMetadata;
 
 @Mapper
 public interface UserMapper {
 
     @Select("SELECT * FROM user WHERE user_id = #{userId}")
-    User getUser(int userId);
+    User getUser(Long userId);
 
     @Select("SELECT * FROM user_metadata WHERE user_id = #{userId}")
-    UserMetadata getUserMetadata(int userId);
+    UserMetadata getUserMetadata(Long userId);
 
     @Select("SELECT user.*, user_metadata.* FROM user JOIN  user_metadata ON user.user_id = user_metadata.user_id WHERE user.user_id = #{userId}")
-    UserDetail getUserDetail(int userId);
+    UserDetail getUserDetail(Long userId);
 
     @Select("SELECT * FROM user")
     List<User> getAllUserList();
